Fix callback function read_data_handler
Curl 8.7.1 has added length checking in order to detect when the announced
length does not match what the input delivers. This change has resulted in an
issue when using the curb gem. This commit fixes the condition check for
remaining vs read_bytes and removes the unneeded branch, and fixes the return
value as the result.